Consolidated content script
